+++ cube/Demo/SimpleDelegate/waterDelegate.h
@@ -24,6 +24,7 @@ private:
     void removeBlock();
 private:
     Camera camera;
+    Camera cameraFixed;
     Scene * scene;
     QVector2D mousePressPosition;
     QVector2D mouseLastPosition;