+++ cube/shader/shaderpoll.cpp
@@ -54,4 +54,5 @@ ShaderPool::ShaderPool()
     this->createShader ("linear_blur","./res/shaders/postProcess.vs","./res/shaders/linearBlur.fs");
     this->createShader ("basic_sprite","./res/shaders/basic_sprite.vs","./res/shaders/basic_sprite.fs");
     this->createShader ("deferred_water","./res/shaders/deferred_water.vs","./res/shaders/deferred_water.fs");
+    this->createShader ("deferred_water_PG","./res/shaders/deferred_water_PG.vs","./res/shaders/deferred_water_PG.fs");
 }