+++ res/shaders/deferred.fs
@@ -27,7 +27,7 @@ vec3 CalcBumpedNormal()
     vec3 Tangent = normalize(v_tangent);
     Tangent = normalize(Tangent - dot(Tangent, Normal) * Normal);
     vec3 Bitangent = cross(Tangent, Normal);
-    vec3 BumpMapNormal = texture2D(g_diffuse_texture, v_texcoord).xyz;
+    vec3 BumpMapNormal = texture2D(g_normal_map, v_texcoord).xyz;
     BumpMapNormal = 2.0 * BumpMapNormal - vec3(1.0, 1.0, 1.0);
     vec3 NewNormal;
     mat3 TBN = mat3(Tangent, Bitangent, Normal);